+++ chat/src/main/java/star/astro/chat/controller/ChatroomController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import star.astro.chat.service.UserService;
+import star.astro.chat.service.ChatroomService;
 
 import java.util.Map;
 
@@ -13,19 +13,19 @@ import java.util.Map;
 public class ChatroomController {
 
     @Autowired
-    private UserService userService;
+    private ChatroomService chatroomService;
 
     @PostMapping("")
     public ResponseEntity<?> createChatroom(@RequestParam Map<String, Object> params) {
         String username = (String) params.get("username");
         String chatroomName = (String) params.get("chatroomName");
-        userService.createChatroom(username, chatroomName);
+        chatroomService.createChatroom(username, chatroomName);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PostMapping("/{chatroomId}/users/{username}")
-    public ResponseEntity<?> joinChatroom(@PathVariable String chatroomId, @PathVariable String username) {
-        userService.joinChatroom(username, chatroomId);
+    public ResponseEntity<?> addUser(@PathVariable String chatroomId, @PathVariable String username) {
+        chatroomService.addUser(username, chatroomId);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
